requirements.txt
Django
djangorestframework
django-debug-toolbar
django-autoslug
psycopg
pytest
flake8
Pillow
colorthief